Azure Boards action (#47)
* Adds a new action to create an Azure Boards work item.

* Fixes work item command

* boards: don't use `$` in env var errors

Generally speaking, when complaining about an missing environment
variable, it is not reported with the leading `$`.  Omit it.

* boards: parse title and description from the event

GitHub Actions writes the event information to a file specified by
`GITHUB_EVENT_PATH`.  Parse that file (using `jq`) for the issue title
and issue body that can be provided to Azure Boards.

* boards: update Azure DevOps URL environment variable

* boards: remove TITLE and DESCRIPTION vars from README

* boards: quote the url

* boards: only execute on issues opened

* boards: thens stay on the same line

* boards: ensure GITHUB_EVENT_PATH is set

Ensure that GITHUB_EVENT_PATH is set, it's required.

* boards: handle reopened and closed work items

Update the work item state to correspond with the GitHub issue state.

* boards: provide sensible defaults

Don't require AZURE_BOARDS_FEATURE, or the updated state names.  Use
sensible defaults of Feature, and the corresponding states for them.

* boards: add issue comments to work item discussion

* Updates work item command to use "System.Tags" in place of the cryptic "80"

* Remove variables from doc that are not required.

* Adds link to source issue when creating work items.

* Adds quotes to hyperlink

* Removes function for link creation.

* Changes closed state from "Done" to "Closed"

* Updated README

* Addresses review feedback.

- Fixes quoting issues.
- Simplifies for loops.
- Readme links to docs for clarity on types and states.
- Cleans up some style issues.

* boards: handle GitHub issue labels

If the `ISSUE_LABEL` variable is set, we will only pay attention to
GitHub issues that have the given label.

* boards: install the discount markdown parser

* boards: parse the description as markdown

GitHub supports markdown in issue descriptions; Azure Boards uses raw
HTML.  Parse the markdown and produce that HTML.

* boards: parse github comments as markdown

* boards: handle extended markdown

* boards: remove some debugging

* Adds support for tagging with repo.full_name to avoid collisions.

* Removes double negative