Azure DevOps configuration updates (#49)
* pipelines: stylistic changes in env var tests

Minor stylistic changes in the environment variable tests: first, in
error messages, do not include the leading `$`.  It's uncommon to
include in error messages and is shell-dependent (PowerShell, eg,
uses a different syntax).  Removal standardizes the messages with the
other scripts in this repository.

Also, do not keep a trailing semicolon on the `if` conditional, which
delimits commands.  The semicolon should either be removed if the `then`
is on a newline, or the `then` should be on the same line as the `if`
(to make the semicolon functional).  I've moved the `then` up to the
line with the `if` to standardize with the other scripts in this
repository.

* release pipelines: stylistic changes in env var tests

Similar to f7cc92ba6dafb764370b094824ca0533182440ee, do not prefix error
messages about environment variables with a `$`.  Similarly, do not keep
an unnecessary trailing semicolon on the `if` conditionals.

* boards: use AZURE_DEVOPS_ variables for configuration

Instead of using `AZURE_BOARDS_` variables for the common configuration
to Azure DevOps, prefix the names with `AZURE_DEVOPS_`.  This allows
users to re-use configuration between multiple instances of GitHub
Actions that target different Azure DevOps services.

This is especially useful for personal access tokens, which are secret
variables and thus not copy/pasteable between GitHub Actions
configurations, but *do* default to be shared.  If the shared secret
variable is `AZURE_DEVOPS_TOKEN` then both boards *and* pipelines
actions can use it without additional configuration or a second token.

* pipelines: use AZURE_DEVOPS_ variables for configuration

Instead of using `AZURE_PIPELINE_` variables for the common
configuration to Azure DevOps, prefix the names with `AZURE_DEVOPS_`.
This allows users to re-use configuration between multiple instances
of GitHub Actions that target different Azure DevOps services.

* release pipelines: use AZURE_DEVOPS_ variables for configuration

Instead of using `AZURE_PIPELINE_` variables for the common
configuration to Azure DevOps, prefix the names with `AZURE_DEVOPS_`.
This allows users to re-use configuration between multiple instances
of GitHub Actions that target different Azure DevOps services.

* Azure DevOps: allow fully qualified URL

Allow users to set `AZURE_DEVOPS_URL` to the actual URL endpoint that
they want to connect to.  This benefits on-premises Azure DevOps Server
users, who would have a bespoke URL on their own network, not a
`dev.azure.com` URL.

Fallback for backward compatibility to continue allowing users to
specify `AZURE_DEVOPS_ORGANIZATION` and creating an Azure DevOps URL
from that.

* azure devops: require that the URL be set

Don't allow AZURE_DEVOPS_ORGANIZATION, simply require a proper URL to
the Azure DevOps account whether that's hosted or on-prem.

* releasepipelines: send error messages to stderr

* azure devops: update READMEs

Update READMEs to reflect the removal of AZURE_DEVOPS_ORGANIZATION